humble bundle add version. next work is Direct Games!
Former-commit-id: 7d6e67e819d5723d58d2f51cb5787865c69fce46

+++ models/sale_info.sql
@@ -7,7 +7,7 @@ create table sale_info(
     start_date datetime,
     end_date datetime,
     sale_percentage int not null,
-    sale_price int not null,
+    sale_price decimal(10,2) not null,
     FOREIGN KEY (game_id) REFERENCES game (id),
     FOREIGN KEY (site_id) REFERENCES site (id)
 );
